% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2R.R
\name{F2R}
\alias{F2R}
\title{Convert an F-test with 1 degree of freedom to effect size r.}
\usage{
F2R(Fstat, N, width = 0.95, neg = F)
}
\arguments{
\item{Fstat}{The F-value of a test}

\item{N}{The total sample size}

\item{width}{Width of confidence interval}

\item{neg}{Indicates when effect size is in direction such that r < 0}
}
\description{
Convert an F-test with 1 degree of freedom to effect size r.
}
