% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_p.R
\name{fix_p}
\alias{fix_p}
\title{Fix a p-value to APA style}
\usage{
fix_p(p)
}
\arguments{
\item{p}{The p-value to be printed}
}
\description{
This function drops the leading 0 and truncates low values to "p < .001"

import relevant namespace
}
