% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{fuse}
\alias{fuse}
\title{Fuse cell means and pool SDs to turn a many cells into two cells}
\usage{
fuse(m1, sd1, n1, m2, sd2, n2)
}
\arguments{
\item{m1}{Vector of means to pool into Group 1}

\item{sd1}{Vector of SDs to pool into Group 1}

\item{m2}{Vector of means to pool into Group 2}

\item{sd2}{Vector of SDs to pool into Group 2}

\item{n1i}{Vector of cell ns to pool into Group 1}

\item{n2i}{Vector of cell ns to pool into Group 2}
}
\description{
Often you need to calculate an effect size from a 2×2 (or more complex!) study design.
Use this function to combine the relevant cells into just two cells for entry into your spreadsheet.
}
\examples{
fuse(m1 = c(2.9, 3.8),
    m2 = c(4.9, 2.8),
    sd1 = c(1.3, 1.8),
    sd2 = c(0.7, 1.3),
    n1 = c(25, 25),
    n2 = c(25, 25))
}
