% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numformat.R
\name{numformat}
\alias{numformat}
\title{Format a number bounded on -1, 1 for APA style}
\usage{
numformat(val)
}
\arguments{
\item{val}{The number to be formatted}
}
\description{
Common targets for this function are p-values, correlations, R^2, etc.
}
