% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_etasq.R
\name{partial_etasq}
\alias{partial_etasq}
\title{Report partial eta-squared given sums of squares}
\usage{
partial_etasq(model, effect, type = 3)
}
\arguments{
\item{model}{GLM object from which to extract SS and SSE}

\item{effect}{Parameter of interest}
}
\description{
Report partial eta-squared given sums of squares
}
