% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_sd.R
\name{pool.sd}
\alias{pool.sd}
\title{Pool standard deviations}
\usage{
pool.sd(sds, ns)
}
\arguments{
\item{sds}{Vector of standard deviations}

\item{ns}{Vector of sample sizes}
}
\description{
This function takes in a vector of standard deviations and sample sizes
to create a pooled standard deviation

import relevant namespace
}
