% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.CI.R
\name{r.CI}
\alias{r.CI}
\title{Fetch confidence interval given a Pearson r.}
\usage{
r.CI(r.equiv, N)
}
\arguments{
\item{r.equiv}{Pearson r}

\item{N}{Total sample size}
}
\description{
This method uses the Fisher's Z to make the CI.
}
