% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_f.R
\name{report_f}
\alias{report_f}
\title{Report an F-test}
\usage{
report_f(model, effect)
}
\arguments{
\item{model}{Model object from which to extract F-value.}

\item{effect}{Parameter with F-value of interest}
}
\description{
Report an F-test as F(df1, df2) = F, p = .###.

import relevant namespace
}
