% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_kt.R
\name{report_kt}
\alias{report_kt}
\title{Report the Kruskal test of medians}
\usage{
report_kt(model)
}
\arguments{
\item{model}{Model object from which to extract chisq-value.}
}
\description{
Report the results of Kruskal-Wallis Rank Sum test as performed by kruskal.test()

import relevant namespace
}
