% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_LR.R
\name{report_LR}
\alias{report_LR}
\title{Report manually-made likelihood-ratio chisq-test}
\usage{
report_LR(row)
}
\arguments{
\item{model}{Model object from which to extract chisq-value.}

\item{effect}{Parameter with chisq-value of interest}
}
\description{
Report an LR chisq-test as chisq(df) = chisq, p = .###.
I am considering deprecating this and replacing it with a generic
reporter function for manually-supplied chi-sqs or stats in general.

import relevant namespace
}
