% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_LRchisq.R
\name{report_LRchisq}
\alias{report_LRchisq}
\title{Report a likelihood-ratio chisq-test}
\usage{
report_LRchisq(model, effect)
}
\arguments{
\item{model}{Model object from which to extract chisq-value.}

\item{effect}{Parameter with chisq-value of interest}
}
\description{
Appropriate for use with "glm" objects.
Report an LR chisq-test as chisq(df) = chisq, p = .###.
}
