% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t.lmer.R
\name{report_t.lmer}
\alias{report_t.lmer}
\title{Report a t-test from an lmer model}
\usage{
report_t.lmer(model, effect)
}
\arguments{
\item{model}{lmer model object from which to extract t-value.}

\item{effect}{Parameter with t-value of interest}
}
\description{
Note that I have no idea what df to supply pt() and that Anova() probably
uses chi-square LR tests for good reason... so this may be a bad idea.
}
\details{
Reports a t-test from an lmer object as t = t, p = ###.
}
