% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_z.R
\name{report_z}
\alias{report_z}
\title{Report a z-test from a logistic regression model}
\usage{
report_z(model, effect)
}
\arguments{
\item{model}{Model object from which to extract F-value.}

\item{effect}{Parameter with F-value of interest}
}
\description{
Report a z-test as z = z, p = .###, OR = OR [OR.LL, OR.UL].
}
