% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2R.R
\name{t2R}
\alias{t2R}
\title{Convert a t-test to effect size r}
\usage{
t2R(tstat, N)
}
\arguments{
\item{tstat}{t-statistic to be used}

\item{N}{total sample size}
}
\description{
Convert a t-test to effect size r
}
