% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_etasq.R
\name{partial_etasq}
\alias{partial_etasq}
\title{Report partial eta-squared given sums of squares}
\usage{
partial_etasq(SS, SSE, df.1, df.2, conf.level = 0.95, digits = 2)
}
\arguments{
\item{SS}{Effect's associated sum of squares.}

\item{SSE}{Model's sum of squared errors.}

\item{df.1}{Degrees of freedom of the effect.}

\item{df.2}{Degrees of freedom of the error term.}

\item{conf.level}{Confidence level of CI. Defaults to .95.}

\item{digits}{Significant digits to print in output. Defaults to 2.}
}
\description{
Report partial eta-squared given sums of squares
}
