% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_kt.R
\name{report_kt}
\alias{report_kt}
\title{Report the kt test}
\usage{
report_kt(model)
}
\arguments{
\item{model}{Model object from which to extract chisq-value.}
}
\description{
I don't remember what this was for -- some kind of goodness-of-fit test?
TODO: Figure that out.
}
