% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t.R
\name{report_t}
\alias{report_t}
\title{Report a t-value from a regression model}
\usage{
report_t(model, effect, metric = "b", n1, n2)
}
\arguments{
\item{model}{Model object from which to extract t-value.}

\item{effect}{Parameter with t-value of interest}
}
\description{
Reports t, p-value, and ESCI on b for a model term.
}
