% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_t.R
\name{report_t}
\alias{report_t}
\title{Report a t-value from a regression model}
\usage{
report_t(model, effect)
}
\arguments{
\item{model}{Model object from which to extract t-value.}

\item{effect}{Parameter with t-value of interest}
}
\description{
Reports t, p-value, and ESCI on r for a model term.
TODO: Figure out if this is bivariate or partial r.
TODO: Report degrees of freedom.

import relevant namespace
}
